+++ src/zandronum/src/g_doom/a_lostsoul.cpp
@@ -49,8 +49,8 @@ void A_SkullAttack(AActor *self, fixed_t speed)
 	an = self->angle >> ANGLETOFINESHIFT;
 	self->velx = FixedMul (speed, finecosine[an]);
 	self->vely = FixedMul (speed, finesine[an]);
-	dist = P_AproxDistance (dest->x - self->x, dest->y - self->y);
-	dist = dist / speed;
+	dist = (int)(P_AproxDistance (dest->x - self->x, dest->y - self->y));
+	dist = (int)(dist / speed);
 	
 	if (dist < 1)
 		dist = 1;
