+++ src/zandronum/src/v_video.cpp
@@ -1748,7 +1748,7 @@ const int BaseRatioSizes[5][4] =
 	{ 1280, 450, 0,                   48*3/4 },		// 16:9   426.6667, 150,      multiplied by three
 	{ 1152, 500, 0,                   48*5/6 },		// 16:10  386,      166.6667, multiplied by three
 	{ 1224, 471, 0,                   48*40/51 },	// 17:10  408,		156.8627, multiplied by three
-	{  960, 640, (int)(6.5*FRACUNIT), 48*15/16 }	//  5:4   320,      213.3333, multiplied by three
+	{  960, 640, (int)(6.5*double(FRACUNIT)), 48*15/16 }	//  5:4   320,      213.3333, multiplied by three
 };
 
 void IVideo::DumpAdapters ()
